apps/api/pyproject.toml
[project]
name = "spotflow-api"
version = "0.1.0"
description = "SpotFlow API for water sports spots"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "alembic",
    "psycopg2-binary",
    "pydantic",
    "python-multipart",
    "python-dotenv",
    "supabase>=2.0.0",
    "geoalchemy2>=0.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]
[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"